setup mac tests
