+++ agent/install_linux.go
@@ -1,5 +1,5 @@
-//go:build !windows
-// +build !windows
+//go:build linux
+// +build linux
 
 /*
 Copyright 2022 AmidaWare LLC.
