+++ agent/umi/umi_unix_test.go
@@ -4,7 +4,6 @@
 package umi_test
 
 import (
-	"reflect"
 	"testing"
 
 	"github.com/amidaware/rmmagent/agent/umi"
@@ -15,26 +14,20 @@ func TestGetInfo(t *testing.T) {
 		name string
 		expected map[string]interface{}
 		atLeast int
-		expectedErrors []error
 	}{
 		{
 			name: "Get info",
 			expected: make(map[string]interface{}),
 			atLeast: 1,
-			expectedErrors: []error{},
 		},
 	}
 
 	for _, tt := range testTable {
 		t.Run(tt.name, func(t *testing.T) {
-			result, errs := umi.GetInfo()
+			result, _ := umi.GetInfo()
 			t.Logf("result: (%v)", result)
 			if len(result) < tt.atLeast {
 				t.Errorf("expected at least %d, got %d", tt.atLeast, len(result))
-			}
-
-			if !reflect.DeepEqual(tt.expectedErrors, errs) {
-				t.Errorf("expected (%v), got (%v)", tt.expectedErrors, errs)
 			}
 		})
 	}
